+++ src/scripts/events/background/lifecycleActions.bg.js
@@ -1,14 +1,8 @@
+import {displayModalOnInstall} from "../../units/interface/modals/displayModalOnInstall.bg";
+
 export function setInstalledAction() {
   chrome.runtime.onInstalled.addListener(function (details) {
-    if (details.reason === "install") {
-      chrome.storage.sync.set({displayModalWelcome: true}, function () {
-        console.log("Welcome modal will be displayed");
-      });
-    } else if (details.reason === "update") {
-      chrome.storage.sync.set({displayModalUpdate: true}, function () {
-        console.log("Last update modal will be displayed");
-      });
-    }
+    displayModalOnInstall(details);
   });
 }
