+++ src/scripts/process/background.js
@@ -2,7 +2,7 @@ import {handleIconClick} from "../units/icon/handleIconClick.bg";
 // import {exportAllThreadsListener} from "../events/exportAll.bg";
 import {clickCountInit} from "../units/interface/modals/clickCount.all";
 import {handleIconImage} from "../units/icon/handleIconImage.bg";
-import {setLifecycleActions} from "./setupBackground/setLifecycleActions";
+import {setLifecycleActions} from "./tasks/setLifecycleActions.bg";
 
 background();
