+++ src/scripts/process/tasks/launchScrapping.tab.js
