+++ src/scripts/process/tasks/setLifecycleActions.bg.js
